Cast to type before resizing image

+++ fibsem/microscope.py
@@ -2924,13 +2924,13 @@ class ThermoMicroscope(FibsemMicroscope):
         resized_points = np.empty((*new_shape, 2), dtype=object)
 
         resized_points[:, :, 0] = transform.resize(
-            points[:, :, 0].reshape(points.shape[0], points.shape[1]),
+            points[:, :, 0].reshape(points.shape[0], points.shape[1]).astype(np.float_),
             output_shape=new_shape,
             order=order,
             preserve_range=True,
         ).astype(np.float_)
         resized_points[:, :, 1] = transform.resize(
-            points[:, :, 1].reshape(points.shape[0], points.shape[1]),
+            points[:, :, 1].reshape(points.shape[0], points.shape[1]).astype(np.uint8),
             output_shape=new_shape,
             order=0,
             preserve_range=True,
